move verify access to concern for code climate
