+++ app/controllers/hearings_controller.rb
@@ -1,10 +1,7 @@
 # frozen_string_literal: true
 
 class HearingsController < ApplicationController
-  before_action :verify_access, except: [:show_print, :show, :update, :find_closest_hearing_locations]
-  before_action :verify_access_to_reader_or_hearings, only: [:show_print, :show]
-  before_action :verify_access_to_hearing_prep_or_schedule, only: [:update]
-  before_action :check_hearing_prep_out_of_service
+  include HearingsConcerns::VerifyAccess
 
   def show
     render json: hearing.to_hash(current_user.id)
@@ -78,18 +75,6 @@ class HearingsController < ApplicationController
 
   def hearing_external_id
     params[:id]
-  end
-
-  def verify_access
-    verify_authorized_roles("Hearing Prep")
-  end
-
-  def verify_access_to_reader_or_hearings
-    verify_authorized_roles("Reader", "Hearing Prep", "Edit HearSched", "Build HearSched")
-  end
-
-  def verify_access_to_hearing_prep_or_schedule
-    verify_authorized_roles("Hearing Prep", "Edit HearSched", "Build HearSched", "RO ViewHearSched")
   end
 
   def set_application
